vector_db/requirements.txt
langchain
langchain-community
langchain-core
faiss-cpu
pypdf
python-docx
pandas
openpyxl
tiktoken
sentence-transformers
pyyaml
chromadb
